requirements.txt
streamlit
pandas
beautifulsoup4
openpyxl
python-docx
fpdf

